fix(pull): extract the pilot predicates into a leaf module (#1766 lint + import fix)
Two problems, one root cause.

`tests/lint_sys_modules.py` failed on a bare `sys.modules.pop(...)` in the new
test's bootstrap — copied from test_capacity_manager.py, but this file imports
every backend module lazily so it never needed the preamble. Removed rather than
papered over with a _STUBBED_MODULE_NAMES declaration: test_1081_physical_meter.py
warns that popping `utils` actively unbinds conftest's canonical importlib
registration, so the idiom is harmful here, not merely unnecessary.

Removing it exposed the real defect. `services/agent_service/__init__.py` eagerly
imports helpers/lifecycle/crud/deploy/terminal, so the guard's
`from services.agent_service.pull_mode import is_pull_pilot_agent` inside
`drain_next` dragged the whole agent-lifecycle stack (→ models) into
backlog_service. Under a test that stubs models/database that surfaced as
`ImportError: cannot import name 'AgentGitConfig' from '<unknown module name>'`
— naming nothing to do with the actual dependency. Verified against clean dev:
tests/unit/test_backlog.py passes 34 there and failed 7 on this branch, so this
was introduced here, not pre-existing.

The predicates now live in `services/pull_pilot.py` — stdlib-only, no package
side effects. `pull_mode` re-exports them so crud/lifecycle and the existing
#1081 tests are untouched; the three dispatch-path callers (capacity_manager,
backlog_service, routers/internal) import the leaf directly. This also retires
a latent instance of the same hazard: capacity_manager's pre-existing
shadow-meter imports had the identical shape and were simply never exercised by
a stubbed test.

A capacity/backlog module should not need the agent CRUD stack to answer "is
this name in an env var".

test_backlog isolation 34 passed; lint clean; combined pull/capacity/canary
suites 284 passed. NOTE: the full unit suite is order-sensitive on dev
independently of this branch — clean dev fails 2 in test_1081_physical_meter
under full-suite ordering while passing in isolation. Both files pass in
isolation and together here; deferring to CI's seeded base/head runs.
